Move Api::Commands to its logical file name
