+++ lib/aruba/api.rb
@@ -6,7 +6,7 @@ require 'aruba/extensions/string/strip'
 
 require 'aruba/platform'
 require 'aruba/api/core'
-require 'aruba/api/command'
+require 'aruba/api/commands'
 
 require 'aruba/api/environment'
 require 'aruba/api/filesystem'
