+++ lib/aruba/api/commands.rb
